Time: 0 ms (100.00%), Space: 5.9 MB (41.61%) - LeetHub

+++ 509-fibonacci-number/509-fibonacci-number.cpp
@@ -1,9 +1,11 @@
 class Solution {
 public:
+        
     int fib(int n) {
       // int f[n+1];
         if(n<2) return n;
-        int a=0,b=1,c;
+        // if n==1 then answer 1, if n==0 f[n]=0
+        long long a=0,b=1,c;
         while(n>1){
             c=a+b;
             a=b;
